Create 1822 Sign of the Product of an Array
Success
Details 
Runtime: 72 ms, faster than 64.03% of Python3 online submissions for Sign of the Product of an Array.
Memory Usage: 14.1 MB, less than 37.74% of Python3 online submissions for Sign of the Product of an Array.

There is a function signFunc(x) that returns:

1 if x is positive.
-1 if x is negative.
0 if x is equal to 0.
You are given an integer array nums. Let product be the product of all values in the array nums.

Return signFunc(product).