Create 1232 Check If It Is a Straight Line
Success
Details 
Runtime: 110 ms, faster than 19.45% of Python3 online submissions for Check If It Is a Straight Line.
Memory Usage: 14.5 MB, less than 28.35% of Python3 online submissions for Check If It Is a Straight Line.

You are given an array coordinates, coordinates[i] = [x, y], where [x, y] represents the coordinate of a point. Check if these points make a straight line in the XY plane.