Create 1678 Goal Parser Interpretation
Success
Details 
Runtime: 35 ms, faster than 71.05% of Python3 online submissions for Goal Parser Interpretation.
Memory Usage: 13.8 MB, less than 96.73% of Python3 online submissions for Goal Parser Interpretation.

You own a Goal Parser that can interpret a string command. The command consists of an alphabet of "G", "()" and/or "(al)" in some order. The Goal Parser will interpret "G" as the string "G", "()" as the string "o", and "(al)" as the string "al". The interpreted strings are then concatenated in the original order.

Given the string command, return the Goal Parser's interpretation of command.